Load entire extension in test script
Eliminate one off bootstrapping of scripts used for testing
in favor of loading the entrie cartodb extension. Previously,
extension scripts were loaded into the test db on an ad hoc
basis. This approach works in some cases but does not handle
dependencies and applies certain functions incorrectly under
the publix schema. This ensures that the latest dev build of
the extension is always available in the test environment.