Create utility to transactionally replace table contents
In use cases such as synchronizations, it becomes necessary
to transactionally replace the contents of a user table.
Previously, this was handled in the editor with a DROP/RENAME
of the old version of the table and a temporary imported
table with the new contents, respectively. This method breaks
down when the table to be synchronized has dependent objects,
such as views which reference it, and the table cannot be
dropped.  In such cases, a TRUNCATE/INSERT is performed in
order to leave the table's dependents intact.