Support parallel CDB_TableMetadata trigger updates
The previous implementation of both deletions from
and updates to CDB_TableMetadata is vulnerable to
deadlock due to the undefined order in which clients
may aquire locks on rows to update or delete.  The
PostgreSQL documentation recommends the FOR UPDATE
statement to solve this issue:

https://www.postgresql.org/docs/current/explicit-locking.html#LOCKING-ROWS

Additionally, the procedure for insertion of records
into CDB_TableMetadata is also modified as part of
this change.  Since multiple clients can generate the
same rows to insert without acuiring locks on those
rows, there is an additional check added to ensure
no rows already present in the table are inserted.

An additional deadlock condition, in which the update
trigger as well as the truncate statement, both seem
to require conflicting locks on the relation being
truncated. The actual content of the trigger, however,
should not depend on the table itself.  Thus the logic
is factored into CDB_TableMetadataTouch and called
directly by CDB_TableUtils_ReplaceTableContents.